Update server.py
I've attempted to add some features, such as calculating and displaying RTT (round-trip time) and breaking the logic into two methods, threaded and serve - this allows the primary serverSocket to be closed and a new one to be generated with each use of a thread, rather than risking the same serverSocket being abused or never properly closed. It still isn't displaying the files in a browser, although all indications are that it should. I remain stumped. (Note: in a client I have written, the files seem to appear just fine. Just not in the browser itself...)